Refactor ConfigEditor
