+++ src/pages/configuration/index.ts
@@ -1,2 +1,2 @@
-export {default as ConfigEditor} from './editor/ConfigEditorWrapper';
+export {default as ConfigEditor} from './editor/editor-wrapper';
 export {default as ConfigList} from './config-list/config-list';
